Fixed some wall collision issue and pellets can now be eaten.

+++ src/pellet.cc
@@ -5,7 +5,7 @@ int Pellet::id_counter = 0;
 Pellet::Pellet(){
 }
 
-Pellet::Pellet(int x, int y, bool eaten):
+Pellet::Pellet(float x, float y, bool eaten):
     id(id_counter++), x(x), y(y), eaten(eaten){
         
     }
@@ -13,10 +13,10 @@ Pellet::~Pellet(){
     
 }
 
-int Pellet::getX(){
+float Pellet::getX(){
     return x + (25/2);
 }
 
-int Pellet::getY(){
+float Pellet::getY(){
     return y + (25/2);
 }
